Assert default values when omitted from acts_as_enum
Columns with default values are allowed to be omitted from an acts_as_enum
specification. These were left up to the database to initialize when
upserting the rows. This meant that if a enum member was changed from a
non-default value to default by removing the attribute declaration then
the table contents would not be updated to the new default value.